backend/requirements.txt
Flask
Flask-Cors
SpeechRecognition
pydub
spacy
python-dotenv
gunicorn
requests